detail

+++ include/nn/swkbd/swkbd.h
@@ -174,7 +174,7 @@ public:
 
 private:
     char* strBuf;
-    s32 bufsize;
+    size_t bufsize;
 };
 
 struct UserWord;   // TODO contents missing
